Created source and asset directories and moved the appropriate files into them. Grouping similar files into folders is a wise thing to do as it makes for a much cleaner project structure.
