+++ assets/ship.bmp
